GODRIVER-318: add canary builders

+++ bson/constructor.go
@@ -65,7 +65,7 @@ func (ElementConstructor) Interface(key string, value interface{}) *Element {
 		switch {
 		case t < math.MaxInt32:
 			elem = EC.Int32(key, int32(t))
-		case t > math.MaxInt64:
+		case uint64(t) > math.MaxInt64:
 			elem = EC.Null(key)
 		default:
 			elem = EC.Int64(key, int64(t))
@@ -129,7 +129,7 @@ func (c ElementConstructor) InterfaceErr(key string, value interface{}) (*Elemen
 		switch {
 		case t < math.MaxInt32:
 			elem = EC.Int32(key, int32(t))
-		case t > math.MaxInt64:
+		case uint64(t) > math.MaxInt64:
 			err = fmt.Errorf("BSON only has signed integer types and %d overflows an int64", t)
 		default:
 			elem = EC.Int64(key, int64(t))
@@ -138,7 +138,7 @@ func (c ElementConstructor) InterfaceErr(key string, value interface{}) (*Elemen
 		switch {
 		case t < math.MaxInt32:
 			elem = EC.Int32(key, int32(t))
-		case t > math.MaxInt64:
+		case uint64(t) > math.MaxInt64:
 			err = fmt.Errorf("BSON only has signed integer types and %d overflows an int64", t)
 		default:
 			elem = EC.Int64(key, int64(t))
